Project refactoring
* Created LuaSamples project to deal with Lua stuff;
* Changed folder references on RPC client and server projects;
* Added BufferedStream to RPCClient instead using NetworkStream, the default buffer size is 5Kb;
* Changed RPCClient object used on screen streaming routine on LuaEngine, to be per LuaEngine instance, this was unsyncronizing video on screen when stream a=was changed to BufferedStream (probably was causing a silent undesired "unknown" behavior on NetworkStream too)
