+++ src/LuaSamples/Source/open_ai/open_ai.lua
@@ -1,0 +1,12 @@
+clear()
+
+print('OpenAI - Chat GPT\n')
+
+question = get_args()
+
+print('Pergunta ==> '..question)
+
+response = ask_gpt(question)
+print('Resposta ==> '..response)
+
+print()
